requirements.txt
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
blessed==1.20.0
boto3==1.28.42
botocore==1.31.42
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-picklefield==3.1
django-redis==5.3.0
django-storages==1.14
djangorestframework==3.12.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
qiniu==7.10.0
redis==3.5.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
